+++ Runtime/CustomAttributes/ArtificeIgnoreAttribute.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace ArtificeToolkit.Attributes
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public class ArtificeIgnoreAttribute : CustomAttribute
+    {
+    }
+}